datapath: Add support for L7 load balancers
Add new SVC_FLAG_L7LOADBALANCER to designate services that are not
loadbalanced in BPF, but are forwarded to a local TPROXY loadbalancer
instead.

The set of proxy ports registered to datapath was limited to the
preconfigured set of HTTP, DNS, and proxylib ports. Allow adding (CRD)
proxy ports to this set.

Redirect L7 LB nodeport service connections received by the host to an
Envoy listener.

Redirect to nodeport L7 LB from bpf_sock for hostns sources. This is
needed due to this case not being handled by eBPF datapath on TC
level. Note that this requires a recent kernel to work (Linux
5.12+). With older kernels nodeport services only work between nodes,
not from a node host namespace to the node itself.

Signed-off-by: Jarno Rajahalme <jarno@isovalent.com>