bpf: lxc: clarify that lb4_rev_nat() in from-container is only for loopback
Per-packet RevDNAT for in-cluster access to services typically happens in
the to-container path, as replies flow back to the client.

There is one exception though: in the loopback case, replies are already
RevDNATed in the from-container path, as they exit the backend. Note that
we have no other case where from-container would match a packet as CT_REPLY
or CT_RELATED to a CT entry with `ct_state->rev_nat_index` set.

Make this clear by adding a check for `ct_state->loopback` in
handle_ipv4_from_lxc(), and outright removing the currently unused code in
the IPv6 function.

Signed-off-by: Julian Wiedmann <jwi@isovalent.com>