WIP bpf: dsr: fix DSR-Geneve without BPF Host Routing
When local delivery uses tailcall, we would usually bypass the stack.
But reply traffic for DSR-Geneve typically *doesn't* pass back into the
tunnel. Instead it gets forwarded to an egress interface. If that forward
happens via the stack, iptables would drop us. So we need to pass through
the stack in forward direction.

TODO the condition in from-overlay should be
     deliver_via_stack := is_dsr && !is_defined(ENABLE_HOST_ROUTING)

Signed-off-by: Julian Wiedmann <jwi@isovalent.com>